+++ Source/Middleware/ZyGames.Framework.Game/Contract/Action/AppStoreDetailAction.cs
@@ -40,7 +40,7 @@ namespace ZyGames.Framework.Game.Contract.Action
 		/// </summary>
 		/// <param name="aActionId">A action identifier.</param>
 		/// <param name="httpGet">Http get.</param>
-        public AppStoreDetailAction(short aActionId, ActionGetter httpGet)
+        public AppStoreDetailAction(int aActionId, ActionGetter httpGet)
             : base(aActionId, httpGet)
         {
         }
